+++ packages/groq/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/groq
 
+## 4.0.0-canary.49
+
+### Patch Changes
+
+- 023550e: Deprecate `streamText` result `fullStream` in favor of `stream`.
+
 ## 4.0.0-canary.48
 
 ### Patch Changes