+++ packages/rsc/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/rsc
 
+## 3.0.0-canary.158
+
+### Patch Changes
+
+- ai@7.0.0-canary.157
+
+## 3.0.0-canary.157
+
+### Patch Changes
+
+- Updated dependencies [023550e]
+- Updated dependencies [e92fc45]
+  - ai@7.0.0-canary.156
+
 ## 3.0.0-canary.156
 
 ### Patch Changes