+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,20 @@
 
 ### Patch Changes
 
+- ai@7.0.0-canary.157
+
+## 0.0.1-canary.0
+
+### Patch Changes
+
+- Updated dependencies [023550e]
+- Updated dependencies [e92fc45]
+  - ai@7.0.0-canary.156
+
+## 0.0.1-canary.0
+
+### Patch Changes
+
 - Updated dependencies [e67d80e]
 - Updated dependencies [6cca112]
 - Updated dependencies [82fc0ab]