+++ packages/workflow/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/workflow
 
+## 1.0.0-canary.74
+
+### Patch Changes
+
+- ai@7.0.0-canary.157
+
+## 1.0.0-canary.73
+
+### Patch Changes
+
+- Updated dependencies [023550e]
+- Updated dependencies [e92fc45]
+  - ai@7.0.0-canary.156
+
 ## 1.0.0-canary.72
 
 ### Patch Changes